agent/requirements.txt
strands-agents[otel]
bedrock-agentcore
